FIx some C++ errors
Clang compiler was used to find more errors than displayed by GCC:
- Default params have to be declared in declaration of a method, not its definition
- fclose can't be called on a NULL FILE* variable

Declaration of CumulatedHistogram was also wrong: base class was declared with double template type

+++ ImageIn/Histogram.tpp
@@ -61,7 +61,7 @@ imagein::CumulatedHistogram::CumulatedHistogram(const imagein::Image_t<D>& img, 
 }
 
 template <typename D>
-imagein::CumulatedHistogram::CumulatedHistogram(const imagein::Image_t<D>& img, const imagein::Rectangle& rect) : imagein::Array<unsigned int>(1 << (sizeof(D)*8))
+imagein::CumulatedHistogram::CumulatedHistogram(const imagein::Image_t<D>& img, const imagein::Rectangle& rect) : imagein::Array<double>(1 << (sizeof(D)*8))
 {
     this->computeHistogram(img, 0, rect);
 }
